+++ .eslintignore
@@ -14,3 +14,5 @@ coverage
 debug
 buildsystem-config.ts
 settings.js
+
+*
