+++ debug/launch/main.ts
@@ -8,8 +8,9 @@ import { ITestingSettings } from "../../test/load-settings.js";
 // add your debugging imports here and prior to submitting a PR git checkout debug/debug.ts
 // will allow you to keep all your debugging files locally
 // comment out the example
-import { Example } from "./sp.js";
+// import { Example } from "./sp.js";
 // import { Example } from "./graph.js";
+import { Example } from "./filter-builder.js";
 
 // setup the connection to SharePoint using the settings file, you can
 // override any of the values as you want here, just be sure not to commit
